chore: Move console to separate package
