+++ pkg/cli/console/console.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package streams
+package console
 
 import (
 	"io"
